fix: corregir tipo incompatible en PremioCajaDTO (Integer vs Double)

+++ src/main/java/com/subastashop/backend/dto/PremioCajaDTO.java
@@ -7,6 +7,6 @@ public class PremioCajaDTO {
     private Integer id;
     private String nombre;
     private String imagenUrl;
-    private Double probabilidad;
+    private Integer probabilidad;
     private Integer stock;
 }
